Implement `HelpObject` trait and migrate help functions to modules
