+++ pacwrap/src/help/config.rs
@@ -19,36 +19,39 @@
 
 use std::fmt::{Result, Write};
 
-use crate::help::{version_string, HelpLayout};
+use crate::help::{version_string, HelpLayout, HelpObject};
 
-pub fn default(buf: &mut String, layout: &HelpLayout) -> Result {
-    let head = layout.head();
-    let tab = layout.tab();
-    let bold = layout.bold();
-    let sub_bold = layout.sub_bold();
-    let sub_para = layout.sub_paragraph();
-    let reset = layout.reset();
-    let reset_bold = layout.reset_bold();
-    let code = layout.code();
-    let name = env!("CARGO_PKG_NAME");
-    let date = env!("PACWRAP_BUILDSTAMP");
+pub struct PacwrapYml;
 
-    match layout {
-        HelpLayout::Man => writeln!(
-            buf,
-            ".nh\n.TH {name}.yml 2 \"{date}\" \"{name} version_string_placeholder\" \"Pacwrap Configuration Directives\"\n"
-        )?,
-        HelpLayout::Markdown => writeln!(
-            buf,
-            "# Pacwrap Configuration Directives
+impl HelpObject for PacwrapYml {
+    fn topic(buf: &mut String, layout: &HelpLayout) -> Result {
+        let head = layout.head();
+        let tab = layout.tab();
+        let bold = layout.bold();
+        let sub_bold = layout.sub_bold();
+        let sub_para = layout.sub_paragraph();
+        let reset = layout.reset();
+        let reset_bold = layout.reset_bold();
+        let code = layout.code();
+        let name = env!("CARGO_PKG_NAME");
+        let date = env!("PACWRAP_BUILDSTAMP");
+
+        match layout {
+            HelpLayout::Man => writeln!(
+                buf,
+                ".nh\n.TH {name}.yml 2 \"{date}\" \"{name} version_string_placeholder\" \"Pacwrap Configuration Directives\"\n"
+            )?,
+            HelpLayout::Markdown => writeln!(
+                buf,
+                "# Pacwrap Configuration Directives
 
 This document was generated by the {name} binary with version {} of the program.\n",
-            version_string(),
-        )?,
-        _ => (),
-    }
+                version_string(),
+            )?,
+            _ => (),
+        }
 
-    writeln!(
+        writeln!(
         buf,
         "{head}NAME{reset}
 {sub_bold}pacwrap.yml{reset_bold} - pacwrap configuration file.
@@ -161,4 +164,5 @@ This document was generated by the {name} binary with version {} of the program.
 {head}COPYRIGHT{reset}
 {tab}This program may be freely redistributed under the terms of the GNU General Public License v3 only.\n"
     )
+    }
 }
