+++ pacwrap/src/help/manual/default.rs
@@ -1,0 +1,109 @@
+/*
+ * pacwrap
+ *
+ * Copyright (C) 2023-2025 Xavier Moffett <sapphirus@azorium.net>
+ * SPDX-License-Identifier: GPL-3.0-only
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+use std::fmt::{Result, Write};
+
+use crate::help::{version_string, HelpLayout, HelpObject};
+
+pub struct Default;
+
+impl HelpObject for Default {
+    fn topic(buf: &mut String, layout: &HelpLayout) -> Result {
+        let head = layout.head();
+        let tab = layout.tab();
+        let sub_para = layout.sub_paragraph();
+        let sub_bold = layout.sub_bold();
+        let bold = layout.bold();
+        let reset = layout.reset();
+        let reset_bold = layout.reset_bold();
+
+        header(buf, layout)?;
+        writeln!(
+            buf,
+            "{head}NAME{reset}
+{tab}pacwrap
+
+{head}SYNOPSIS{reset}
+{tab}pacwrap [{bold}OPERATION{reset_bold} | {bold}VERB{reset_bold}] [{bold}ARGUMENTS{reset_bold}] [{bold}TARGETS{reset_bold}]	
+
+{head}DESCRIPTION{reset}
+{sub_para}A package management front-end which utilises libalpm to facilitate the creation of unprivileged, 
+{tab}namespace containers with parallelised, filesystem-agnostic deduplication. These containers
+{tab}are constructed with bubblewrap to execute package transactions and launch applications.
+
+{sub_para}This application is designed to allow for the creation and execution of secure, replicable 
+{tab}containerised environments for general-purpose use. CLI and GUI applications are all supported. 
+{tab}Once a container environment is configured, it can be re-established or replicated on any system. 
+
+{sub_para}Each long-option parameter can also be specified herein as a command verb for a matter of convenience.
+{tab}Additional command verb shortcuts are available and are documented alongside their relevancy.
+
+{head}OPERATIONS{reset}
+{sub_bold}-E, --exec, run{reset_bold}
+{tab}{tab}Invoke a container to execute the provided command sequence.
+
+{sub_bold}-S, --sync{reset_bold}
+{tab}{tab}Synchronize package databases and containers in aggregate. 
+
+{sub_bold}-R, --remove{reset_bold}
+{tab}{tab}Remove packages from target containers in aggregate.
+
+{sub_bold}-C, --compose{reset_bold}
+{tab}{tab}Compose a container from configuration.
+
+{sub_bold}-Q, --query{reset_bold}
+{tab}{tab}Query package information from target container.
+
+{sub_bold}-P, --process{reset_bold}
+{tab}{tab}Manage and show status of running container processes.
+
+{sub_bold}-L, --list{reset_bold}
+{tab}{tab}List available containers managed by pacwrap.
+
+{sub_bold}-d, --desktop{reset_bold}
+{tab}{tab}Manage desktop files associated with pacwrap containers.
+
+{sub_bold}-U, --utils{reset_bold}
+{tab}{tab}Engage miscellaneous utilities to manage containers.
+
+{sub_bold}-V, --version{reset_bold}
+{tab}{tab}Display version banner or information.
+
+{sub_bold}-h, --help{reset_bold} <{bold}OPERATION{reset_bold} | {bold}VERB{reset_bold} | {bold}TOPIC{reset_bold}>
+{tab}{tab}Print the help manual to {bold}STDOUT{reset_bold}.\n"
+        )
+    }
+}
+
+fn header(buf: &mut String, layout: &HelpLayout) -> Result {
+    let name = env!("CARGO_PKG_NAME");
+    let date = env!("PACWRAP_BUILDSTAMP");
+
+    match layout {
+        HelpLayout::Man => writeln!(buf, ".nh\n.TH {name} 1 \"{date}\" \"{name} version_string_placeholder\" \"User Manual\"\n"),
+        HelpLayout::Markdown => writeln!(
+            buf,
+            "# Pacwrap User Manual
+
+This document was generated by the {name} binary with version {} of the program.\n",
+            version_string()
+        ),
+        _ => Ok(()),
+    }
+}
